+++ app/src/main/java/com/kiral/charityapp/domain/fake/FakeProfile.kt
@@ -4,7 +4,7 @@ data class FakeProfile(
     val id: Int,
     val name: String,
     var region: String,
-    var credit: Float,
+    var credit: Double,
     val email: String,
     //var categories ??
     //var Birthday
@@ -15,14 +15,14 @@ var fakeProfiles = mutableListOf(
         id = 0,
         name = "Alžbeta Pekná",
         region = "cze",
-        credit = 150f,
+        credit = 150.0,
         email = "alzbeta@email.com"
     ),
     FakeProfile(
         id = 1,
         name = "Ivan Horváth",
         region = "svk",
-        credit = 100f,
+        credit = 100.0,
         email = "ivan@email.com"
     )
 )
